line_chatbot_api.py
from linebot import (
    LineBotApi, WebhookHandler
)
from linebot.exceptions import (
    InvalidSignatureError
)
from linebot.models import (
    MessageEvent, 
    PostbackEvent,
    TextMessage, 
    TextSendMessage, 
    ImageSendMessage, 
    StickerSendMessage, 
    LocationSendMessage,
    TemplateSendMessage,
    ButtonsTemplate,
    PostbackAction,
    MessageAction,
    URIAction,
    CarouselTemplate,
    CarouselColumn,
    ImageCarouselTemplate,
    ImageCarouselColumn,
    DatetimePickerAction,
    ConfirmTemplate,
    QuickReply,
    QuickReplyButton
)

line_bot_api = LineBotApi('YYvAi7/NhGIm+qAJsACII1WPQ1vxA95V04l433DCmXgNW68KOAimhcjtoBtKhdJ4IlUNaRVUBT2VlWpIiTn2J9v6CMpvpQX0woW8DlvIsINuH8H12+xdEfpl/S3yEKnpAPGdiO2D25+87BV5P1XEOQdB04t89/1O/w1cDnyilFU=')
handler = WebhookHandler('007151bd77914be37695c86d07265c82')