requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
future==0.18.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
requests==2.26.0
SpeechRecognition==3.8.1
urllib3==1.26.7
Werkzeug==2.0.2
